1487 Move XML read to separate xml/components/read file
